created hook for searchMentionUser and showCommands
